requirements.txt
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.6
django-import-export==1.0.1
et-xmlfile==1.0.1
gunicorn==19.8.1
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.4
psycopg2==2.7.4
pytz==2018.4
PyYAML==3.12
requests==2.19.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
